+++ packages/@kagal-ed25519-secret/src/index.ts
@@ -26,6 +26,7 @@ export {
 } from './key-record';
 export {
   type KeyConfig,
+  newSecret,
   parseSecretsToKeys,
   parseSecretToKey,
 } from './secret';
@@ -41,12 +42,14 @@ export {
 export {
   asBytes,
   asMessageBytes,
+  atLeast,
   type Bytes,
   decodeASCII,
   decodeBase64,
   encodeBase64,
   encodeKey,
   getRandom,
+  isInRange,
   splitFirst,
   splitLast,
 } from './utils';